+++ InterInfoBackScript/src/bean/CustomsMonthBean.java
@@ -4,7 +4,7 @@ public class CustomsMonthBean {
 	private String border;
 	private String year;
 	private String month;
-	private int MONTO;
+	private double MONTO;
 	
 	public String getBorder() {
 		return border;
@@ -18,7 +18,7 @@ public class CustomsMonthBean {
 	public String getMonth() {
 		return month;
 	}
-	public int getMONTO() {
+	public double getMONTO() {
 		return MONTO;
 	}
 
@@ -28,7 +28,7 @@ public class CustomsMonthBean {
 	public void setMonth(String month) {
 		this.month = month;
 	}
-	public void setMONTO(int mONTO) {
+	public void setMONTO(double mONTO) {
 		MONTO = mONTO;
 	}
