+++ InterInfoBackScript/src/bean/CustomsQuarterBean.java
@@ -5,7 +5,7 @@ public class CustomsQuarterBean {
 	private String border;
 	private String year;
 	private String month;
-	private int MONTO;
+	private double MONTO;
 	public String getBorder() {
 		return border;
 	}
@@ -15,7 +15,7 @@ public class CustomsQuarterBean {
 	public String getMonth() {
 		return month;
 	}
-	public int getMONTO() {
+	public double getMONTO() {
 		return MONTO;
 	}
 	public void setBorder(String border) {
@@ -27,7 +27,7 @@ public class CustomsQuarterBean {
 	public void setMonth(String month) {
 		this.month = month;
 	}
-	public void setMONTO(int mONTO) {
+	public void setMONTO(double mONTO) {
 		MONTO = mONTO;
 	}
