+++ InterInfoBackScript/src/bean/HSBeanSemister.java
@@ -4,7 +4,7 @@ public class HSBeanSemister {
 	private String FRACCION;
 	private String year;
 	private String month;
-	private int MONTO;
+	private double MONTO;
 	public String getFRACCION() {
 		return FRACCION;
 	}
@@ -14,7 +14,7 @@ public class HSBeanSemister {
 	public String getMonth() {
 		return month;
 	}
-	public int getMONTO() {
+	public double getMONTO() {
 		return MONTO;
 	}
 	public void setFRACCION(String fRACCION) {
@@ -26,7 +26,7 @@ public class HSBeanSemister {
 	public void setMonth(String month) {
 		this.month = month;
 	}
-	public void setMONTO(int mONTO) {
+	public void setMONTO(double mONTO) {
 		MONTO = mONTO;
 	}
